# hooks/api/gen/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package hookapigen

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn20AllowedHeaders = map[string]string{
		"GET": "Last-Event-Id",
	}
	rn15AllowedHeaders = map[string]string{
		"PATCH": "Content-Type",
	}
	rn16AllowedHeaders = map[string]string{
		"PATCH": "Content-Type",
	}
	rn13AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "changes"

				if l := len("changes"); len(elem) >= l && elem[0:l] == "changes" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleHooksListChangesRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'e': // Prefix: "e"

				if l := len("e"); len(elem) >= l && elem[0:l] == "e" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'v': // Prefix: "vents"

					if l := len("vents"); len(elem) >= l && elem[0:l] == "vents" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleHooksListEventsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/stream"

						if l := len("/stream"); len(elem) >= l && elem[0:l] == "/stream" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleHooksStreamEventsRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: rn20AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'x': // Prefix: "xecution"

					if l := len("xecution"); len(elem) >= l && elem[0:l] == "xecution" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "PATCH":
							s.handleHooksSetExecutionRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "PATCH",
								allowedHeaders: rn15AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "application/json",
							})
						}

						return
					}

				}

			case 'h': // Prefix: "hooks"

				if l := len("hooks"); len(elem) >= l && elem[0:l] == "hooks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleHooksListRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "hookId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'e': // Prefix: "execution"

							if l := len("execution"); len(elem) >= l && elem[0:l] == "execution" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "PATCH":
									s.handleHooksSetHookExecutionRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "PATCH",
										allowedHeaders: rn16AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "application/json",
									})
								}

								return
							}

						case 'o': // Prefix: "output"

							if l := len("output"); len(elem) >= l && elem[0:l] == "output" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleHooksOutputRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "GET",
										allowedHeaders: nil,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						case 'r': // Prefix: "run"

							if l := len("run"); len(elem) >= l && elem[0:l] == "run" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleHooksRunHookRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn13AllowedHeaders,
										acceptPost:     "application/json",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					}

				}

			case 'p': // Prefix: "ping"

				if l := len("ping"); len(elem) >= l && elem[0:l] == "ping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleHooksPingRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'q': // Prefix: "queue"

				if l := len("queue"); len(elem) >= l && elem[0:l] == "queue" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleHooksListQueueRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'r': // Prefix: "runs"

				if l := len("runs"); len(elem) >= l && elem[0:l] == "runs" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleHooksListRunsRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'h': // Prefix: "hutdown"

					if l := len("hutdown"); len(elem) >= l && elem[0:l] == "hutdown" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleHooksShutdownRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'n': // Prefix: "napshots"

					if l := len("napshots"); len(elem) >= l && elem[0:l] == "napshots" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleHooksListSnapshotsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 't': // Prefix: "tatus"

					if l := len("tatus"); len(elem) >= l && elem[0:l] == "tatus" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleHooksStatusRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'w': // Prefix: "wait"

				if l := len("wait"); len(elem) >= l && elem[0:l] == "wait" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleHooksWaitRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "changes"

				if l := len("changes"); len(elem) >= l && elem[0:l] == "changes" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = HooksListChangesOperation
						r.summary = "List observed file changes."
						r.operationID = "hooks-list-changes"
						r.operationGroup = ""
						r.pathPattern = "/changes"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'e': // Prefix: "e"

				if l := len("e"); len(elem) >= l && elem[0:l] == "e" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'v': // Prefix: "vents"

					if l := len("vents"); len(elem) >= l && elem[0:l] == "vents" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = HooksListEventsOperation
							r.summary = "List recent hook daemon audit events."
							r.operationID = "hooks-list-events"
							r.operationGroup = ""
							r.pathPattern = "/events"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/stream"

						if l := len("/stream"); len(elem) >= l && elem[0:l] == "/stream" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = HooksStreamEventsOperation
								r.summary = "Stream hook daemon audit events as Server-Sent Events."
								r.operationID = "hooks-stream-events"
								r.operationGroup = ""
								r.pathPattern = "/events/stream"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'x': // Prefix: "xecution"

					if l := len("xecution"); len(elem) >= l && elem[0:l] == "xecution" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "PATCH":
							r.name = HooksSetExecutionOperation
							r.summary = "Pause or resume all hook execution."
							r.operationID = "hooks-set-execution"
							r.operationGroup = ""
							r.pathPattern = "/execution"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'h': // Prefix: "hooks"

				if l := len("hooks"); len(elem) >= l && elem[0:l] == "hooks" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = HooksListOperation
						r.summary = "List discovered hooks with current status metadata."
						r.operationID = "hooks-list"
						r.operationGroup = ""
						r.pathPattern = "/hooks"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "hookId"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'e': // Prefix: "execution"

							if l := len("execution"); len(elem) >= l && elem[0:l] == "execution" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "PATCH":
									r.name = HooksSetHookExecutionOperation
									r.summary = "Pause or resume one hook."
									r.operationID = "hooks-set-hook-execution"
									r.operationGroup = ""
									r.pathPattern = "/hooks/{hookId}/execution"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'o': // Prefix: "output"

							if l := len("output"); len(elem) >= l && elem[0:l] == "output" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = HooksOutputOperation
									r.summary = "Return captured output for a hook's latest run."
									r.operationID = "hooks-output"
									r.operationGroup = ""
									r.pathPattern = "/hooks/{hookId}/output"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'r': // Prefix: "run"

							if l := len("run"); len(elem) >= l && elem[0:l] == "run" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = HooksRunHookOperation
									r.summary = "Request a hook run."
									r.operationID = "hooks-run-hook"
									r.operationGroup = ""
									r.pathPattern = "/hooks/{hookId}/run"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						}

					}

				}

			case 'p': // Prefix: "ping"

				if l := len("ping"); len(elem) >= l && elem[0:l] == "ping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = HooksPingOperation
						r.summary = "Check daemon reachability and metadata."
						r.operationID = "hooks-ping"
						r.operationGroup = ""
						r.pathPattern = "/ping"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'q': // Prefix: "queue"

				if l := len("queue"); len(elem) >= l && elem[0:l] == "queue" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = HooksListQueueOperation
						r.summary = "List queued hook work."
						r.operationID = "hooks-list-queue"
						r.operationGroup = ""
						r.pathPattern = "/queue"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'r': // Prefix: "runs"

				if l := len("runs"); len(elem) >= l && elem[0:l] == "runs" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = HooksListRunsOperation
						r.summary = "List hook run history."
						r.operationID = "hooks-list-runs"
						r.operationGroup = ""
						r.pathPattern = "/runs"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'h': // Prefix: "hutdown"

					if l := len("hutdown"); len(elem) >= l && elem[0:l] == "hutdown" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = HooksShutdownOperation
							r.summary = "Request graceful daemon shutdown."
							r.operationID = "hooks-shutdown"
							r.operationGroup = ""
							r.pathPattern = "/shutdown"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'n': // Prefix: "napshots"

					if l := len("napshots"); len(elem) >= l && elem[0:l] == "napshots" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = HooksListSnapshotsOperation
							r.summary = "List workspace snapshots."
							r.operationID = "hooks-list-snapshots"
							r.operationGroup = ""
							r.pathPattern = "/snapshots"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 't': // Prefix: "tatus"

					if l := len("tatus"); len(elem) >= l && elem[0:l] == "tatus" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = HooksStatusOperation
							r.summary = "Return daemon and hook status summary."
							r.operationID = "hooks-status"
							r.operationGroup = ""
							r.pathPattern = "/status"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'w': // Prefix: "wait"

				if l := len("wait"); len(elem) >= l && elem[0:l] == "wait" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = HooksWaitOperation
						r.summary = "Wait for daemon hook work and watcher processing to settle."
						r.operationID = "hooks-wait"
						r.operationGroup = ""
						r.pathPattern = "/wait"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}