# hooks/api/openapi/hooks.yaml
openapi: 3.1.0
info:
  title: Discobox Hooks Daemon API
  version: 0.1.0
  description: Session-scoped hooks daemon API served over HTTP on a Unix domain socket.
servers:
  - url: http://unix
paths:
  /ping:
    get:
      operationId: hooks-ping
      summary: Check daemon reachability and metadata.
      responses:
        '200':
          description: Daemon metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
  /status:
    get:
      operationId: hooks-status
      summary: Return daemon and hook status summary.
      responses:
        '200':
          description: Daemon status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /wait:
    get:
      operationId: hooks-wait
      summary: Wait for daemon hook work and watcher processing to settle.
      parameters:
        - name: timeout_seconds
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Final settled or timed-out daemon status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitResponse'
  /hooks:
    get:
      operationId: hooks-list
      summary: List discovered hooks with current status metadata.
      responses:
        '200':
          description: Hook list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HooksResponse'
  /events:
    get:
      operationId: hooks-list-events
      summary: List recent hook daemon audit events.
      parameters:
        - $ref: '#/components/parameters/HookIDQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Audit event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
  /events/stream:
    get:
      operationId: hooks-stream-events
      summary: Stream hook daemon audit events as Server-Sent Events.
      parameters:
        - $ref: '#/components/parameters/HookIDQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: SSE stream carrying one Event JSON object in each data field.
          content:
            text/event-stream:
              schema:
                type: string
  /runs:
    get:
      operationId: hooks-list-runs
      summary: List hook run history.
      parameters:
        - $ref: '#/components/parameters/HookIDQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Hook run history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunsResponse'
  /changes:
    get:
      operationId: hooks-list-changes
      summary: List observed file changes.
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Observed file changes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangesResponse'
  /snapshots:
    get:
      operationId: hooks-list-snapshots
      summary: List workspace snapshots.
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Workspace snapshots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsResponse'
  /queue:
    get:
      operationId: hooks-list-queue
      summary: List queued hook work.
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Queued hook work.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueResponse'
  /execution:
    patch:
      operationId: hooks-set-execution
      summary: Pause or resume all hook execution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionPatchRequest'
      responses:
        '200':
          description: Updated global execution state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
  /shutdown:
    post:
      operationId: hooks-shutdown
      summary: Request graceful daemon shutdown.
      responses:
        '200':
          description: Shutdown accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShutdownResponse'
  /hooks/{hookId}/execution:
    patch:
      operationId: hooks-set-hook-execution
      summary: Pause or resume one hook.
      parameters:
        - $ref: '#/components/parameters/HookIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionPatchRequest'
      responses:
        '200':
          description: Updated hook execution state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /hooks/{hookId}/run:
    post:
      operationId: hooks-run-hook
      summary: Request a hook run.
      parameters:
        - $ref: '#/components/parameters/HookIDPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '200':
          description: Hook run request result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /hooks/{hookId}/output:
    get:
      operationId: hooks-output
      summary: Return captured output for a hook's latest run.
      parameters:
        - $ref: '#/components/parameters/HookIDPath'
      responses:
        '200':
          description: Captured output.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputResponse'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    HookIDPath:
      name: hookId
      in: path
      required: true
      schema:
        type: string
    HookIDQuery:
      name: hook_id
      in: query
      required: false
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PingResponse:
      $ref: ./components/transport.yaml#/PingResponse
    StatusResponse:
      $ref: ./components/transport.yaml#/StatusResponse
    WaitResponse:
      $ref: ./components/transport.yaml#/WaitResponse
    HooksResponse:
      $ref: ./components/transport.yaml#/HooksResponse
    EventsResponse:
      $ref: ./components/transport.yaml#/EventsResponse
    RunsResponse:
      $ref: ./components/transport.yaml#/RunsResponse
    ChangesResponse:
      $ref: ./components/transport.yaml#/ChangesResponse
    SnapshotsResponse:
      $ref: ./components/transport.yaml#/SnapshotsResponse
    QueueResponse:
      $ref: ./components/transport.yaml#/QueueResponse
    ExecutionPatchRequest:
      $ref: ./components/transport.yaml#/ExecutionPatchRequest
    ExecutionResponse:
      $ref: ./components/transport.yaml#/ExecutionResponse
    RunRequest:
      $ref: ./components/transport.yaml#/RunRequest
    RunResponse:
      $ref: ./components/transport.yaml#/RunResponse
    OutputResponse:
      $ref: ./components/transport.yaml#/OutputResponse
    ShutdownResponse:
      $ref: ./components/transport.yaml#/ShutdownResponse
    ErrorResponse:
      $ref: ./components/transport.yaml#/ErrorResponse
    HookStatus:
      $ref: ./components/model/core.yaml#/HookStatus
    Hook:
      $ref: ./components/model/core.yaml#/Hook
    HookType:
      $ref: ./components/model/core.yaml#/HookType
    HookEngine:
      $ref: ./components/model/core.yaml#/HookEngine
    RunAs:
      $ref: ./components/model/core.yaml#/RunAs
    Status:
      $ref: ./components/model/core.yaml#/Status
    Event:
      $ref: ./components/model/core.yaml#/Event
    ChangedFile:
      $ref: ./components/transport.yaml#/ChangedFile
    Run:
      $ref: ./components/transport.yaml#/Run
    ObservedFileChange:
      $ref: ./components/transport.yaml#/ObservedFileChange
    SnapshotOmission:
      $ref: ./components/transport.yaml#/SnapshotOmission
    WorkspaceSnapshot:
      $ref: ./components/transport.yaml#/WorkspaceSnapshot
    QueuedHook:
      $ref: ./components/transport.yaml#/QueuedHook